pyproject.toml
[tool.poetry]
name = "embedfinder"
version = "0.1.0"
description = ""
authors = ["Luke100000 <jonas.peche@aon.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.14"
numpy = "<2"
transformers = "^4.43.3"
pillow = "^10.4.0"
pathspec = "^0.12.1"
chromadb = "^0.5.18"
pyqt5 = "^5.15.11"
librosa = "^0.10.2.post1"
diskcache = "^5.6.3"
torch = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
